feature.  https://gitlab.iptco.me/tio/wip/-/issues/99 获取大额输赢数据

+++ module_hall/src/main/java/com/walisport/module/hall/data/GameAllRankingListData.kt
@@ -3,7 +3,7 @@ package com.walisport.module.hall.data
 data class GameAllRankingListData(
     val gameIconUrl: String ,
     val gameName: String ,
-    val multiple : Float ,
+    val multiple: Int ,
     val countryIcon: Int ,
     val symbol: String ,
     val result: Float ,
